pyproject.toml
[tool.poetry]
name = "counter-strike-scrapers"
version = "0.1.0"
description = ""
authors = ["Stelios Bitzas <stelios.bitzas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
bs4 = "^0.0.2"
playwright = "^1.42.0"
flask = "^3.0.2"
requests = "^2.31.0"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^4.0.0"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pre-commit = "^3.7.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
